refactor: consolidate import statements and enhance logging in security modules for improved clarity and consistency
